+++ dev/test.sh
@@ -8,3 +8,8 @@ npx start-server-and-test --expect 404 \
   "npx @stoplight/prism-cli mock -d --json-schema-faker-fillProperties=false dev/specs/ai-guard.openapi.json" \
   4010 \
   "go test ./pangea-sdk/service/ai_guard/mock_test.go"
+
+npx start-server-and-test --expect 404 \
+  "npx @stoplight/prism-cli mock -d --json-schema-faker-fillProperties=false dev/specs/prompt-guard.openapi.json" \
+  4010 \
+  "go test ./pangea-sdk/service/prompt_guard/mock_test.go"